+++ jest.config.js
@@ -6,7 +6,7 @@ module.exports = {
   ...ts_preset,
   ...dynamo_preset,
   testEnvironment: 'node',
-  testPathIgnorePatterns: ['bin/', 'dist/'],
+  testPathIgnorePatterns: ['bin/', 'dist/', 'cdk.out/'],
   coverageThreshold: {
     global: {
       statements: 80,
