+++ test/unit/handlers/check-order-status.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable */
 import { DutchInput, DutchOutput, OrderValidation, TokenTransfer } from '@uniswap/uniswapx-sdk'
 import { BigNumber } from 'ethers'
-import { ORDER_STATUS } from '../../lib/entities'
-import { CheckOrderStatusHandler } from '../../lib/handlers/check-order-status/handler'
-import { FILL_EVENT_LOOKBACK_BLOCKS_ON, getSettledAmounts } from '../../lib/handlers/check-order-status/util'
-import { NATIVE_ADDRESS } from '../../lib/util/constants'
+import { ORDER_STATUS } from '../../../lib/entities'
+import { CheckOrderStatusHandler } from '../../../lib/handlers/check-order-status/handler'
+import { FILL_EVENT_LOOKBACK_BLOCKS_ON, getSettledAmounts } from '../../../lib/handlers/check-order-status/util'
+import { NATIVE_ADDRESS } from '../../../lib/util/constants'
 import { ORDER_INFO } from '../fixtures'
 import { MOCK_ORDER_ENTITY, MOCK_ORDER_HASH } from './test-data'
