+++ test/unit/handlers/get-docs.test.ts
@@ -1,5 +1,5 @@
-import { GetDocsHandler } from '../../lib/handlers/get-docs/GetDocsHandler'
-import schema from '../../swagger.json'
+import { GetDocsHandler } from '../../../lib/handlers/get-docs/GetDocsHandler'
+import schema from '../../../swagger.json'
 import { HeaderExpectation } from '../utils'
 
 describe('Testing get api docs json handler.', () => {
