+++ test/unit/handlers/get-orders.test.ts
@@ -1,6 +1,6 @@
 import { OrderType } from '@uniswap/uniswapx-sdk'
-import { ORDER_STATUS, SORT_FIELDS } from '../../lib/entities'
-import { GetOrdersHandler } from '../../lib/handlers/get-orders/handler'
+import { ORDER_STATUS, SORT_FIELDS } from '../../../lib/entities'
+import { GetOrdersHandler } from '../../../lib/handlers/get-orders/handler'
 import { HeaderExpectation } from '../utils'
 
 describe('Testing get orders handler.', () => {
