+++ test/unit/handlers/on-chain-status-checker.integration.test.ts
@@ -3,10 +3,10 @@ import { StaticJsonRpcProvider } from '@ethersproject/providers'
 import { EventWatcher, OrderValidation, OrderValidator } from '@uniswap/uniswapx-sdk'
 import { DocumentClient } from 'aws-sdk/clients/dynamodb'
 import { BigNumber } from 'ethers'
-import { BATCH_READ_MAX, OnChainStatusChecker } from '../../lib/compute/on-chain-status-checker'
-import { ORDER_STATUS } from '../../lib/entities'
-import { log } from '../../lib/Logging'
-import { LimitOrdersRepository } from '../../lib/repositories/limit-orders-repository'
+import { BATCH_READ_MAX, OnChainStatusChecker } from '../../../lib/compute/on-chain-status-checker'
+import { ORDER_STATUS } from '../../../lib/entities'
+import { log } from '../../../lib/Logging'
+import { LimitOrdersRepository } from '../../../lib/repositories/limit-orders-repository'
 import { deleteAllRepoEntries } from '../utils'
 import { dynamoConfig, MOCK_ORDER_ENTITY, MOCK_ORDER_HASH } from './test-data'
